Fixes and Organizing
Fixed player alignment on grid

fixed player sprite visibility in between certain tiles in the map
----------
organized game objects involved with the camera under one object

organized the tilemap (Grid object) using game objects grouping common tilemaps.

Made a naming scheme for the tilemaps to understand their purpose better
i.e. F2-U == this tilemap is the second layer of the floor for the map and appears under the player

B3-0-C == this tilemap is the third layer for buildings on the map, appears over the player, and holds collision tiles