fixed some bugs, added db connection conditions, added comments
Fixed some bugs:
bug1: At load screen, null references but game continues to run.
bug1fix: Move scripts to a more accessible location for the scenes.
bug2: UserManager not establishing region progression to the user.
bug2fix: Added variables there and in and RegisterButton.
db connection section entailed mostly establishing a boolean variable in the GameManager and then adding in conditions for the scripts that make saves or load the game, the buttons that click them were also adjusted.